[ 서버 상태 추적 ]
- Heartbeat를 기존 Stream기반에서 hash기반으로 바꿈
- 자동 새로고침 추가
- 유저 수 그래프 추가

+++ WebServer/WebServer/Monitoring/ServerMetrics.cs
@@ -2,10 +2,10 @@
 {
     public static class ServerMetrics
     {
-        private static int _requestCount = 0;
+        private static long _requestCount = 0;
         private static int _onlineUsers = 0;
 
-        public static int RequestCount => _requestCount;
+        public static long RequestCount => _requestCount;
         public static int OnlineUserCount => _onlineUsers;
 
         public static void IncrementRequest() => Interlocked.Increment(ref _requestCount);
